backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23

# MySQL Support
PyMySQL==1.1.0
cryptography==41.0.7

# AI & API
groq>=0.36.0
python-dotenv==1.0.0

# Utilities
requests==2.31.0
